+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.4.4
 * support vue2.7 setup macros `defineProps` `defineEmits` `defineExpose` `withDefaults`
+* upgrade dependencies `eslint-plugin-vue` `vue-eslint-parser`
 ## 1.4.3
 * remove: eslint.rc `'spaced-comment': 1` when use typescript.
 * add: use `vue-eslint-parser` parser in (`vue2` / `vue3`) eslint config file